requirements.txt
ultralytics
transformers
numpy
datasets
Pillow
torch
opencv-python
onnx
onnxruntime
fastapi
sqlalchemy
alembic
pydantic
firebase-admin
openai>=1.0.0